tidy up file formats
