+++ adm/style/acp_postlove.html
@@ -13,7 +13,7 @@
 					<span>{L_POSTLOVE_SHOW_LIKES_EXPLAIN}</span>
 				</dt>
 				<dd>
-					<label><input type="radio" id="postlove_show_likes" name="poslove[postlove_show_likes]" value="1"<!-- IF POST_LIKES --> checked<!-- ENDIF -->>{L_YES}</label><label><input type="radio" id="postlove_show_likes" name="poslove[postlove_show_likes]" value="0"<!-- IF !POST_LIKES --> checked<!-- ENDIF -->>{L_NO}</label>
+					<label><input type="radio" id="postlove_show_likes" name="poslove[postlove_show_likes]" value="1"<!-- IF POST_LIKES --> checked<!-- ENDIF -->>{L_YES}</label><label><input type="radio" id="postlove_show_likes" name="poslove[postlove_show_likes]"	value="0"<!--	IF	!POST_LIKES	-->	checked<!--	ENDIF	-->>{L_NO}</label>
 				</dd>
 			</dl>
 			<dl>
@@ -22,7 +22,7 @@
 					<span>{L_POSTLOVE_SHOW_LIKED_EXPLAIN}</span>
 				</dt>
 				<dd>
-					<label><input type="radio" id="postlove_show_liked" name="poslove[postlove_show_liked]" value="1"<!-- IF POST_LIKED --> checked<!-- ENDIF -->>{L_YES}</label><label><input type="radio" id="postlove_show_liked" name="poslove[postlove_show_liked]" value="0"<!-- IF !POST_LIKED --> checked<!-- ENDIF -->>{L_NO}</label>
+					<label><input type="radio" id="postlove_show_liked" name="poslove[postlove_show_liked]" value="1"<!-- IF POST_LIKED --> checked<!-- ENDIF -->>{L_YES}</label><label><input type="radio" id="postlove_show_liked" name="poslove[postlove_show_liked]"	value="0"<!--	IF	!POST_LIKED	-->	checked<!--	ENDIF	-->>{L_NO}</label>
 				</dd>
 			</dl>
 			<dl>
@@ -31,7 +31,7 @@
 					<span>{L_POSTLOVE_AUTHOR_LIKE_EXPLAIN}</span>
 				</dt>
 				<dd>
-					<label><input type="radio" id="postlove_author_like" name="poslove[postlove_author_like]" value="1"<!-- IF AUTHOR_LIKE --> checked<!-- ENDIF -->>{L_YES}</label><label><input type="radio" id="postlove_author_like" name="poslove[postlove_author_like]" value="0"<!-- IF !AUTHOR_LIKE --> checked<!-- ENDIF -->>{L_NO}</label>
+					<label><input type="radio" id="postlove_author_like" name="poslove[postlove_author_like]" value="1"<!-- IF AUTHOR_LIKE --> checked<!-- ENDIF -->>{L_YES}</label><label><input type="radio" id="postlove_author_like" name="poslove[postlove_author_like]"	value="0"<!--	IF	!AUTHOR_LIKE	-->	checked<!--	ENDIF	-->>{L_NO}</label>
 				</dd>
 			</dl>
 			<dl>
@@ -40,42 +40,42 @@
 					<span>{L_POSTLOVE_SHOW_BUTTON_EXPLAIN}</span>
 				</dt>
 				<dd>
-					<label><input type="radio" id="postlove_show-button" name="poslove[postlove_show_button]" value="1"<!-- IF SHOW_BUTTON --> checked<!-- ENDIF -->>{L_YES}</label><label><input type="radio" id="postlove_show_button" name="poslove[postlove_show_button]" value="0"<!-- IF !SHOW_BUTTON --> checked<!-- ENDIF -->>{L_NO}</label>
+					<label><input type="radio" id="postlove_show-button" name="poslove[postlove_show_button]" value="1"<!-- IF SHOW_BUTTON --> checked<!-- ENDIF -->>{L_YES}</label><label><input type="radio" id="postlove_show_button" name="poslove[postlove_show_button]"	value="0"<!--	IF	!SHOW_BUTTON	-->	checked<!--	ENDIF	-->>{L_NO}</label>
 				</dd>
 			</dl>
-        <table>
-            <tr>
-                   <th><label>{L_POSTLOVE_HOWMANY_MOST_LIKED}</label></th>
-                   <th><label>{L_POSTLOVE_INDEX}</label></th>
-                   <th><label>{L_POSTLOVE_FORUM}</label></th>
-            </tr>
-            <tr>
-                <td>{L_POSTLOVE_HOWMANY_MOST_LIKED_DAY}</td>
-                <td><input type="text" name="poslove[postlove_index_most_liked_today]" size="4" value="{INDEX_HOWMANY_TODAY}" /></td>
-		        <td><input type="text" name="poslove[postlove_forum_most_liked_today]" size="4" value="{FORUM_HOWMANY_TODAY}" /></td>
-            <tr>
-                <td>{L_POSTLOVE_HOWMANY_MOST_LIKED_WEEK}</td>
-                <td><input type="text" name="poslove[postlove_index_most_liked_this_week]" size="4" value="{INDEX_HOWMANY_THIS_WEEK}" /></td>
-		        <td><input type="text" name="poslove[postlove_forum_most_liked_this_week]" size="4" value="{FORUM_HOWMANY_THIS_WEEK}" /></td>
-            <tr>
-                <td>{L_POSTLOVE_HOWMANY_MOST_LIKED_MONTH}</td>
-                <td><input type="text" name="poslove[postlove_index_most_liked_this_month]" size="4" value="{INDEX_HOWMANY_THIS_MONTH}" /></td>
-		        <td><input type="text" name="poslove[postlove_forum_most_liked_this_month]" size="4" value="{FORUM_HOWMANY_THIS_MONTH}" /></td>
-            <tr>
-                <td>{L_POSTLOVE_HOWMANY_MOST_LIKED_YEAR}</td>
-                <td><input type="text" name="poslove[postlove_index_most_liked_this_year]" size="4" value="{INDEX_HOWMANY_THIS_YEAR}" /></td>
-		        <td><input type="text" name="poslove[postlove_forum_most_liked_this_year]" size="4" value="{FORUM_HOWMANY_THIS_YEAR}" /></td>
-            <tr>
-                <td>{L_POSTLOVE_HOWMANY_MOST_LIKED_EVER}</td>
-                <td><input type="text" name="poslove[postlove_index_most_liked_ever]" size="4" value="{INDEX_HOWMANY_EVER}" /></td>
-		        <td><input type="text" name="poslove[postlove_forum_most_liked_ever]" size="4" value="{FORUM_HOWMANY_EVER}" /></td>
-            </tr>
-        </table>
-        </fieldset>
-	    <p class="submit-buttons">
+		<table>
+			<tr>
+				<th><label>{L_POSTLOVE_HOWMANY_MOST_LIKED}</label></th>
+				<th><label>{L_POSTLOVE_INDEX}</label></th>
+				<th><label>{L_POSTLOVE_FORUM}</label></th>
+			</tr>
+			<tr>
+				<td>{L_POSTLOVE_HOWMANY_MOST_LIKED_DAY}</td>
+				<td><input type="text" name="poslove[postlove_index_most_liked_today]" size="4" value="{INDEX_HOWMANY_TODAY}" /></td>
+				<td><input type="text" name="poslove[postlove_forum_most_liked_today]" size="4" value="{FORUM_HOWMANY_TODAY}" /></td>
+			<tr>
+				<td>{L_POSTLOVE_HOWMANY_MOST_LIKED_WEEK}</td>
+				<td><input type="text" name="poslove[postlove_index_most_liked_this_week]" size="4" value="{INDEX_HOWMANY_THIS_WEEK}" /></td>
+				<td><input type="text" name="poslove[postlove_forum_most_liked_this_week]" size="4" value="{FORUM_HOWMANY_THIS_WEEK}" /></td>
+			<tr>
+				<td>{L_POSTLOVE_HOWMANY_MOST_LIKED_MONTH}</td>
+				<td><input type="text" name="poslove[postlove_index_most_liked_this_month]" size="4" value="{INDEX_HOWMANY_THIS_MONTH}" /></td>
+				<td><input type="text" name="poslove[postlove_forum_most_liked_this_month]" size="4" value="{FORUM_HOWMANY_THIS_MONTH}" /></td>
+			<tr>
+				<td>{L_POSTLOVE_HOWMANY_MOST_LIKED_YEAR}</td>
+				<td><input type="text" name="poslove[postlove_index_most_liked_this_year]" size="4" value="{INDEX_HOWMANY_THIS_YEAR}" /></td>
+				<td><input type="text" name="poslove[postlove_forum_most_liked_this_year]" size="4" value="{FORUM_HOWMANY_THIS_YEAR}" /></td>
+			<tr>
+				<td>{L_POSTLOVE_HOWMANY_MOST_LIKED_EVER}</td>
+				<td><input type="text" name="poslove[postlove_index_most_liked_ever]" size="4" value="{INDEX_HOWMANY_EVER}" /></td>
+				<td><input type="text" name="poslove[postlove_forum_most_liked_ever]" size="4" value="{FORUM_HOWMANY_EVER}" /></td>
+			</tr>
+		</table>
+		</fieldset>
+		<p class="submit-buttons">
 			<input class="button1" type="submit" id="submit" name="submit" value="{L_SUBMIT}" />&nbsp;
 			<input class="button2" type="reset" id="reset" name="reset" value="{L_RESET}" />
-			
+
 			<fieldset>
 				<dl>
 					<dt>
