+++ language/en/info_acp_postlove.php
@@ -37,13 +37,13 @@ $lang = array_merge($lang, array(
 
 	'CLEN'	=> 'Clean',
 
-    //Version 2.0 
+	//Version 2.0
 	'POSTLOVE_HOWMANY_MOST_LIKED_DAY'     => 'How many liked-today posts to show',
 	'POSTLOVE_HOWMANY_MOST_LIKED_WEEK'	=> 'How many liked-this-week posts to show',
 	'POSTLOVE_HOWMANY_MOST_LIKED_MONTH'	=> 'How many liked-this-month posts to show',
 	'POSTLOVE_HOWMANY_MOST_LIKED_YEAR'    => 'How many liked-this-year posts to show',
 	'POSTLOVE_HOWMANY_MOST_LIKED_EVER'	=> 'How many liked-ever posts to show',
-	'POSTLOVE_FORUM'    => 'How many to show on Forum pages',
+	'POSTLOVE_FORUM'	=> 'How many to show on Forum pages',
 	'POSTLOVE_INDEX'	=> 'How many to show on Index page',
 	'POSTLOVE_SHOW_BUTTON'	=> 'Show the Post like count in a Post Button?',
 	'POSTLOVE_SHOW_BUTTON_EXPLAIN'	=> 'The Post like count status and action link may be shown as a Post Button at the top of the post or in the old format at the bottom of the post',
