--- styles/prosilver/template/add_bookmark.html
@@ -1,23 +1,0 @@
-<!-- INCLUDE overall_header.html -->
-
-<form method="post" action="{U_POST_ACTION}">
-	<div class="panel">
-		<div class="inner">
-			<p>{L_BOOKMARK_EXPLAIN}</p>
-			<fieldset class="fields1 column1">
-				<dl>
-					<dt><label for="bookmark_desc">{L_COMMENT}{L_COLON}</label></dt>
-					<dd><input name="bookmark_desc" id="bookmark_desc" value="" class="inputbox" type="text" maxlength="40"></dd>
-				</dl>
-			</fieldset>
-
-			<div class="clear"></div>
-			<hr>
-			<fieldset class="submit-buttons">
-				<input value="{L_CANCEL}" name="reset" class="button2" type="submit">&nbsp;
-				<input name="book" value="{L_SUBMIT}" class="button1" type="submit">
-			</fieldset>
-		</div>
-	</div>
-</form>
-<!-- INCLUDE overall_footer.html -->
